fixed Plat
Due to the previous changes where the definition of extra was changed, the time based component was longer by 1 unit, so we need to increase the length of the cohort by one as well.

Suspect that the functions didn't work as intended for skip_extra=false, so updated here too.